+++ src/article/tests/test_manager.py
@@ -1,0 +1,32 @@
+from django.contrib.auth import get_user_model
+from django.test import TestCase
+
+
+class UserManagersTest(TestCase):
+    def test_create_user(self):
+        User = get_user_model()
+        user = User.objects.create_user(email="test@test.com", password="test")
+        self.assertEqual(user.email, "test@test.com")
+        self.assertTrue(user.is_active)
+        self.assertFalse(user.is_staff)
+        self.assertFalse(user.is_superuser)
+        with self.assertRaises(TypeError):
+            User.objects.create_user()
+        with self.assertRaises(TypeError):
+            User.objects.create_user(email="")
+        with self.assertRaises(ValueError):
+            User.objects.create_user(email="", password="test")
+
+    def test_create_superuser(self):
+        User = get_user_model()
+        admin_user = User.objects.create_superuser(
+            email="test@test.com", password="test"
+        )
+        self.assertEqual(admin_user.email, "test@test.com")
+        self.assertTrue(admin_user.is_active)
+        self.assertTrue(admin_user.is_staff)
+        self.assertTrue(admin_user.is_superuser)
+        with self.assertRaises(ValueError):
+            User.objects.create_superuser(
+                email="super@user.com", password="foo", is_superuser=False
+            )
